Migrate some ID types from int to int64 (#100)
* Fix type of DomainRegistration.(ID|DomainID|RegistrantID) to int64
* Fix type of DomainRenewal.(ID|DomainID) to int64
* Fix type of DomainTransfer.(ID|DomainID|RegistrantID) to int64
* Fix type of AccessToken.AccountID to int64

+++ dnsimple/oauth_test.go
@@ -33,7 +33,7 @@ func TestOauthService_ExchangeAuthorizationForToken(t *testing.T) {
 		t.Fatalf("Oauth.ExchangeAuthorizationForToken() returned error: %v", err)
 	}
 
-	want := &AccessToken{Token: "zKQ7OLqF5N1gylcJweA9WodA000BUNJD", Type: "Bearer", AccountID: 1}
+	want := &AccessToken{Token: "zKQ7OLqF5N1gylcJweA9WodA000BUNJD", Type: "Bearer", AccountID: int64(1)}
 	if !reflect.DeepEqual(token, want) {
 		t.Errorf("Oauth.ExchangeAuthorizationForToken() returned %+v, want %+v", token, want)
 	}
